Remove util-functions and dead code, move useful functions to file-utils.ts
