+++ src/types/aliases.d.ts
@@ -1,2 +1,1 @@
-// eslint-disable-next-line import/prefer-default-export
 export type Optional<T> = T | undefined;
